backend/requirements.txt
celery==5.3.6
Django==3.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
redis==3.5.3
psycopg2-binary==2.8.6
psutil==5.9.5
requests==2.32.4
django-redis==5.2.0
hiredis==3.2.1
